insertWorks

+++ app/src/main/java/com/example/naivybeats/models/restaurant/model/Restaurant.java
@@ -2,6 +2,7 @@ package com.example.naivybeats.models.restaurant.model;
 
 import com.example.naivybeats.models.user.model.Users;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.Time;
 import java.time.LocalDateTime;
@@ -16,7 +17,7 @@ public class Restaurant extends Users
     {
     }
 
-    public Restaurant(int user_id, String name, String photo, String email, String password, int phone_number, Date creation_date, Date edition_date, Date deleted_at, int province_id, Double latitud, Double longitud, String direction, Time openingTime, Time closingTime)
+    public Restaurant(int user_id, String name, String photo, String email, String password, int phone_number, Date creation_date, Date edition_date, Date deleted_at, int province_id, BigDecimal latitud, BigDecimal longitud, String direction, Time openingTime, Time closingTime)
     {
         super(user_id, name, photo, email, password, phone_number, creation_date, edition_date, deleted_at, province_id, latitud, longitud);
         this.direction = direction;
